Changed input type from datetime-local to datetime
The way Chrome handles this is totally worthless. I expected it to include to offset alongside the date, e.g., I pick 4pm today and it sends "2013-06-29T16:00:00+0930" to the server. Turns out that Chrome just sends "2013-06-29T16:00:00" which isn't even a valid RFC3339 datetime because of the lack of offset.