Added new prompt voting mode: approval votes
Each user can approve as many prompts as they like, i.e., giving each a score of 1 or 0. The prompt with the highest number of approvals becomes the event's prompt.

This involved changing the way the prompt times are chosen. Instead of being saved into the events table, they're calculated based on the prompt voting type. "Faceoff" voting (the name for the old type) is one hour long, so the voting starts one hour before the event starts. Prompt submissions are allowed for any time between the event's creation and when the prompt voting begins. This also means that the event's "start" is actually *after* the prompt round.