added award_id to theory
This solves the last big issue with the data structure. The last plan was to
create a dummy 'entry" and assign an award to that. This was a pretty bad
solution for obvious reasons.

This forces the scoreboard code to have separate considerations for awards from
entrys and awards from theorys. But that's not a problem.

The alternative is to add a theory_id column to awards. This would allowed a
theory to receive more than one award, which is currently unecessary.