fix t0 to being set globally
Previously, t0 was set in the callback. This meant that, for example, when new
posts were loaded into the page, their t0 was when they were loaded, not when
the page was loaded. Then, when comparing to `now`, which was set on page load,
the resulting time deltas were wrong.