functioning double caching, mediocre results
STILL INSECURE: [% and %] need to be escaped in the nested template.

The general idea is we render a page of posts into one big template.

The big template has already done most of the rendering, with the only
bits left being things like whether the edit button appears.

Currently, compiling a 100 post page takes 3s, and rendering it .1s.
Compared to previous times of .6s/.3s, this is pretty bad.

It seem necessary to use a much more lightweight templating system for the
nested templates to be really useful.