change session store from file to cookie
Issue with the file store is that it's difficult to expire old sessions.
We want sessions to last a long time, but any non-user request will create
a new session taking up more and more disk space. This has lead to the /tmp
directory being filled every ~1.5 months, requiring a system reboot.

It should be possible to promote sessions to expire in 1y after the user
logs in, but then cleaning up the expired sessions is non-trivial. Using
a cookie store solves the problem in a simple way, and since the session
is not storing any significant data, the limitations of a cookie store
aren't an issue.