fix regression that broke guesses
WriteOff::Controller::Guess::guess previously expected $c->stash->{mode}
to be a string, but it was changed to be a WriteOff::Mode object. This
wasn't accounted for when creating the Theory database entry, resulting
in a bunch of mode = WriteOff::Mode=SCALAR(0x...) entries that are never
found again.

Fix is simple, but to prevent similar issues from occuring in future
WriteOff::Mode is also made to overload on stringification to its name.